+++ python/jaeger/actor/commands/__init__.py
@@ -12,7 +12,6 @@ from clu.parsers.click import command_parser
 jaeger_parser = command_parser
 
 
-from .bootloader import *
 from .debug import *
 from .fvc import *
 from .ieb import *