+++ python/jaeger/actor/commands/positioner.py
@@ -6,19 +6,31 @@
 # @Filename: commands.py
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
+from __future__ import annotations
+
 import asyncio
 import pathlib
+
+from typing import TYPE_CHECKING, List, Tuple
 
 import click
 import numpy
 
 import clu
 
-from jaeger.commands import SetCurrent, Trajectory
-from jaeger.exceptions import TrajectoryError
+from jaeger import config
+from jaeger.commands import CommandID, SetCurrent, Trajectory
+from jaeger.exceptions import JaegerError, TrajectoryError
 from jaeger.utils import get_goto_move_time
 
 from . import jaeger_parser
+
+
+if TYPE_CHECKING:
+    from clu import Command
+
+    from jaeger import FPS
+    from jaeger.actor import JaegerActor
 
 
 __all__ = [
@@ -50,67 +62,88 @@ def check_positioners(positioner_ids, command, fps, initialised=False):
 
 
 @jaeger_parser.command()
-@click.argument("POSITIONER-ID", type=int, nargs=-1)
+@click.argument("POSITIONER-IDS", type=int, nargs=-1)
 @click.argument("ALPHA", type=click.FloatRange(-360.0, 360.0))
 @click.argument("BETA", type=click.FloatRange(-360.0, 360.0))
 @click.option(
     "-r",
     "--relative",
     is_flag=True,
-    help="whether this is a relative move",
+    help="Whether this is a relative move",
 )
 @click.option(
     "-s",
     "--speed",
-    type=click.FloatRange(0.0, 2000.0),
+    type=click.FloatRange(100.0, 4000.0),
     nargs=2,
-    help="the speed of both alpha and beta arms, in RPS on the input.",
+    help="The speed of both alpha and beta arms, in RPS on the input.",
 )
 @click.option(
     "-a",
     "--all",
     is_flag=True,
     default=False,
-    help="applies to all valid positioners.",
+    help="Applies to all valid positioners.",
 )
 @click.option(
     "-f",
     "--force",
     is_flag=True,
     default=False,
-    help="forces a move to happen.",
+    help="Forces a move to happen.",
 )
-async def goto(command, fps, positioner_id, alpha, beta, speed, all, force, relative):
+@click.option(
+    "--use-sync/-no-use-sync",
+    " /-S",
+    default=True,
+    help="Whether to use the SYNC line to start the trajectory.",
+)
+async def goto(
+    command: Command[JaegerActor],
+    fps: FPS,
+    positioner_ids: Tuple[int, ...],
+    alpha: float,
+    beta: float,
+    speed: Tuple[float, float],
+    all: bool = False,
+    force: bool = False,
+    relative: bool = False,
+    use_sync: bool = True,
+):
     """Sends positioners to a given (alpha, beta) position."""
+
+    assert command.actor
 
     if all:
         if not force:
-            return command.fail(
-                error="need to specify --force to move all positioners at once."
-            )
-        positioner_id = list(fps.positioners.keys())
+            return command.fail(error="Use --force to move all positioners at once.")
+        pids: List[int] = list(fps.positioners.keys())
+    else:
+        pids = list(positioner_ids)
 
     if not relative:
         if alpha < 0 or beta < 0:
-            return command.fail(error="negative angles only allowed in relative mode.")
+            return command.fail(error="Negative angles only allowed in relative mode.")
 
-    if not check_positioners(positioner_id, command, fps, initialised=True):
+    if not check_positioners(pids, command, fps, initialised=True):
         return
 
     if fps.moving:
         return command.fail(error="FPS is moving. Cannot send goto.")
 
-    speed = speed or [None, None]
+    if speed is None:
+        default_speed = config["positioner"]["motor_speed"]
+        speed = (default_speed, default_speed)
+
     max_time = 0.0
 
-    tasks = []
-    for pid in positioner_id:
+    for pid in pids:
 
         # Manually calculate the max move time we'll encounter.
         p_alpha, p_beta = fps[pid].position
 
         if p_alpha is None or p_beta is None:
-            return command.fail(error="some positioners do not know their positions.")
+            return command.fail(error="Some positioners do not know their positions.")
 
         delta_alpha = abs(p_alpha - alpha) if not relative else alpha
         delta_beta = abs(p_beta - beta) if not relative else beta
@@ -129,24 +162,21 @@ async def goto(command, fps, positioner_id, alpha, beta, speed, all, force, rela
         if time_beta > max_time:
             max_time = time_beta
 
-        tasks.append(
-            fps.positioners[pid].goto(
-                alpha,
-                beta,
-                speed=speed,
-                relative=relative,
-            )
-        )
-
     command.info(move_time=round(max_time, 2))
 
-    result = await clu.as_complete_failer(tasks, on_fail_callback=fps.stop_trajectory)
+    try:
+        await command.actor.fps.goto(
+            pids,
+            alpha,
+            beta,
+            speed=speed,
+            relative=relative,
+            use_sync_line=use_sync,
+        )
+    except JaegerError as err:
+        return command.fail(error=f"Goto command failed: {err}")
 
-    if not result[0]:
-        error_message = result[1] or "goto command failed"
-        command.set_status(clu.CommandStatus.FAILED, error=error_message)
-    else:
-        command.set_status(clu.CommandStatus.DONE, text="Position reached")
+    command.finish(text="Position reached")
 
 
 @jaeger_parser.command()
@@ -212,7 +242,7 @@ async def initialise(command, fps, positioner_id, datums=False):
 
 
 @jaeger_parser.command()
-@click.argument("POSITIONER-ID", type=int, nargs=-1, required=False)
+@click.argument("POSITIONER-IDS", type=int, nargs=-1, required=False)
 @click.option(
     "-f",
     "--full",
@@ -220,13 +250,23 @@ async def initialise(command, fps, positioner_id, datums=False):
     default=False,
     help="outputs more statuses.",
 )
-async def status(command, fps, positioner_id, full):
+async def status(command, fps, positioner_ids, full):
     """Reports the position and status bit of a list of positioners."""
 
-    positioner_ids = positioner_id or list(fps.positioners.keys())
+    positioner_ids = positioner_ids or list(fps.positioners.keys())
 
-    if not check_positioners(positioner_id, command, fps):
+    if not check_positioners(positioner_ids, command, fps):
         return
+
+    await fps.update_status(positioner_ids=positioner_ids)
+    await fps.update_position(positioner_ids=positioner_ids)
+
+    n_trajs = (
+        await fps.send_command(
+            CommandID.GET_NUMBER_TRAJECTORIES,
+            positioner_ids=positioner_ids,
+        )
+    ).get_replies()
 
     command.info(locked=fps.locked)
 
@@ -238,11 +278,13 @@ async def status(command, fps, positioner_id, full):
         alpha_pos = -999 if p.alpha is None else numpy.round(p.alpha, 4)
         beta_pos = -999 if p.beta is None else numpy.round(p.beta, 4)
 
+        n_trajs_pid = n_trajs[pid] if n_trajs[pid] is not None else "?"
+
         if pid in fps.positioner_to_bus:
             interface, bus = fps.positioner_to_bus[pid]
             interface = fps.can.interfaces.index(interface) + 1
         else:
-            interface = "NA"
+            interface = -1
             bus = -1
 
         command.write(
@@ -257,6 +299,7 @@ async def status(command, fps, positioner_id, full):
                 p.firmware or "?",
                 interface,
                 bus,
+                n_trajs_pid,
             ],
         )
 
@@ -319,13 +362,13 @@ async def unlock(command, fps):
     """Unlocks the FPS."""
 
     if not fps.locked:
-        command.info(locker=False)
+        command.info(locked=False)
         return command.finish(text="FPS is not locked")
 
     result = await fps.unlock()
 
     if result:
-        command.info(locker=False)
+        command.info(locked=False)
         return command.finish(text="FPS unlocked")
     else:
         return command.fail(error="failed to unlock FPS")
@@ -340,7 +383,9 @@ async def trajectory(command, fps, path):
         return command.fail(error="FPS is moving. Cannot send trajectory.")
 
     if fps.locked:
-        return command.fail(error="FPS is locked. Cannot send trajectory.")
+        return command.fail(
+            error=f"FPS is locked by {fps.locked_by}. Cannot send trajectory."
+        )
 
     path = pathlib.Path(path).expanduser()
     if not path.exists():