+++ tests/actor/test_actor.py
@@ -34,7 +34,7 @@ async def test_status(actor):
     assert command.status.did_succeed
 
     # cmd running + locked + n_positioners + 5 positioners + done
-    assert len(actor.mock_replies) == 10
+    assert len(actor.mock_replies) == 12
 
 
 async def test_info(actor):