+++ tests/test_positioner.py
@@ -7,7 +7,6 @@
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
 import asyncio
-import logging
 
 import pytest
 
@@ -43,17 +42,6 @@ async def test_goto(vfps, event_loop):
     assert await vfps[1].goto(1, 1)
 
 
-async def test_goto_no_move(vfps, event_loop, caplog):
-
-    caplog.set_level(logging.INFO)
-
-    await vfps.initialise()
-
-    assert await vfps[1].goto(0, 0)
-
-    assert "did not move" in caplog.records[-1].message
-
-
 async def test_goto_relative(vfps, event_loop):
 
     await vfps.initialise()
@@ -67,7 +55,7 @@ async def test_goto_safe_mode(vfps, monkeypatch):
     await vfps.initialise()
 
     with pytest.raises(PositionerError):
-        await vfps[1].goto(100, 150)
+        await vfps[1].goto(100, 150, use_trajectory=False)
 
 
 async def test_goto_safe_mode_custom_beta(vfps, monkeypatch):
@@ -76,4 +64,4 @@ async def test_goto_safe_mode_custom_beta(vfps, monkeypatch):
     await vfps.initialise()
 
     with pytest.raises(PositionerError):
-        await vfps[1].goto(100, 169)
+        await vfps[1].goto(100, 169, use_trajectory=False)